% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-fl.R
\name{append.fl}
\alias{append.fl}
\title{Append adds an element to the end of the list.}
\usage{
\method{append}{fl}(ls, el)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{el}{A data element (from R's standard data types) to add to the list}

\item{...}{Used for S3 consistency}
}
\value{
A closure with the \code{fl} class added. The closure can be used to access
different properties of the list: either the head, tail or its empty status.
}
\description{
This is the more useful function from a user's perspective, since the
organization of the arguments makes it easy to pipe several calls of \code{append}
together. This function combines recursion and the function \code{\link{prepend}}
to make sure the base \code{\link{empty_list}} stays at the end of the list.
}
\seealso{
\code{\link[base]{append}}
}

