% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{append.default}
\alias{append}
\alias{append.default}
\title{New Append Generic}
\usage{
\method{append}{default}(x, values, after = length(x))

append(...)
}
\arguments{
\item{...}{Arguments for different versions of \code{append}}
}
\description{
So that we can have both the original and new version in our package
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Preserve Original Function
}}
\seealso{
\code{\link[base]{append}}
}

