% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concat}
\alias{concat}
\title{Combine two lists into one}
\usage{
concat(lsa, lsb)
}
\arguments{
\item{lsa}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{lsb}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}
}
\value{
A closure with the \code{fl} class added. The closure can be used to access
different properties of the list: either the head, tail or its empty status.
}
\description{
Combine two lists into one
}
\seealso{
\code{\link[base]{c}}
}

