% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{drop.default}
\alias{drop}
\alias{drop.default}
\title{New Drop Generic}
\usage{
\method{drop}{default}(x)

drop(...)
}
\arguments{
\item{...}{Arguments to various versions of \code{drop}}
}
\description{
So that we can have both the original and new version in our package
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Preserve original version
}}
\seealso{
\code{\link[base]{drop}}
}

