% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{filter.fl}
\alias{filter.fl}
\title{Return values that fulfill a predicate function}
\usage{
\method{filter}{fl}(ls, fn, ...)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{fn}{A predicate function, as described above}

\item{...}{Additional arguments to \code{fn}}
}
\value{
A functional list whose data elements return \code{true} in the predicate
function.
}
\description{
A predicate function is one that takes a single argument returns either true or
false. \code{filter} takes such a function, applies it to a functional list,
and keeps those values that return true in the predicate.
}
\seealso{
\code{\link{remove}}, \code{\link[base]{Filter}}
}

