% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-fl.R
\name{filter.default}
\alias{filter}
\alias{filter.default}
\title{New generic filter to keep both versions}
\usage{
\method{filter}{default}(x, filter, method = c("convolution", "recursive"),
  sides = 2L, circular = FALSE, init = NULL)

filter(...)
}
\arguments{
\item{...}{Arguments to different versions of filter}
}
\description{
New generic filter to keep both versions
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Preserve original function
}}
\seealso{
\code{\link[stats]{filter}}
}

