% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-fl.R
\name{find.fl}
\alias{find.fl}
\title{Return the first object in a list that fulfills a predicate}
\usage{
\method{find}{fl}(ls, fn, ...)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{fn}{A predicate function, as described above}

\item{...}{Additional arguments to remove, for S3 consistency}
}
\value{
Either the first object that fulfills the predicate or \code{FALSE}.
}
\description{
A predicate function is one that takes a single argument returns either true or
false. \code{find} takes such a function, applies it to a functional list,
and returns the first object where the function returns \code{TRUE}.
}
\seealso{
\code{\link[base]{funprog}}
}

