% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-fl.R
\name{find.default}
\alias{find}
\alias{find.default}
\title{New Find Generic}
\usage{
\method{find}{default}(what, mode = "any", numeric = FALSE,
  simple.words = TRUE)

find(...)
}
\arguments{
\item{...}{Arguments to different versions of \code{find}}
}
\description{
So that we can have both the original and new version in our package
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Preserve original function
}}
\seealso{
\code{\link[utils]{find}}
}

