% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-fl.R
\name{head.fl}
\alias{head.fl}
\title{Access the head of a functional list}
\usage{
\method{head}{fl}(x, ...)
}
\arguments{
\item{x}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{...}{Additional arguments to \code{head} (for argument consistency)}
}
\value{
The element stored in the head argument of the functional list
}
\description{
Our first selector. Like \code{\link{tail}} and \code{\link{is_empty}},
it uses an anonymous function that returns a particular component of the functional
list. Here that component is the functional list's \emph{head}.
}

