% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\title{Check whether a list is empty}
\usage{
is_empty(ls)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}
}
\value{
The empty status of a functional list
}
\description{
This function is part of the \link[=lambdaList]{List out of Lambda} set of functions.
Like \code{\link{head}} and \code{\link{tail}}, this uses an anoymous function to
return a particular component of a functional list. Here, that's the empty status
of the functional list.
}

