% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{is_equal}
\alias{is_equal}
\title{Test if two functional lists are equal}
\usage{
is_equal(lsa, lsb)
}
\arguments{
\item{lsa}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{lsb}{A function, with \emph{class} \code{fl}, that operates as a functional
list object. This needs to be the same length as \code{lsa}}
}
\value{
A logical: TRUE or FALSE
}
\description{
Test if two functional lists are equal
}

