% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{map}
\alias{map}
\title{Apply a function to a list}
\usage{
map(ls, fn, ...)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{fn}{A function be applied to each of the scalars contained in the arguments
of the functional list object}

\item{...}{Additional arguments to \code{fn}}
}
\value{
A functional list of the same length as \code{ls}, where each new element
is equal to \code{fn(x)} of each scalar contained in \code{ls}
}
\description{
One of the most obvious extensions of a functional list is a tool to iterate
over that "object."
}
\seealso{
\code{\link[base]{lapply}}, \code{\link[base]{Map}}
}

