% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2.R
\name{map2}
\alias{map2}
\title{Two-list map}
\usage{
map2(lsa, lsb, fn, ...)
}
\arguments{
\item{lsa}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{lsb}{A function, with \emph{class} \code{fl}, that operates as a functional
list object. This needs to be the same length as \code{lsa}}

\item{fn}{A function be applied to each of the scalars contained in the arguments
of the functional list object}

\item{...}{Additional arguments to \code{fn}}
}
\value{
A functional list of the same length as \code{lsa} and \code{lsb}, where each new element
is equal to \code{fn(x)} of each scalar contained in \code{ls}
}
\description{
Apply a two-argument function to two lists. This can be used to implement
vectorized versions of common functions.
}

