% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth.R
\name{nth}
\alias{nth}
\title{Select an element of a functional list}
\usage{
nth(ls, n)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{n}{An integer referring the cardinal position of a scalar in the
functional list object}
}
\value{
A scalar contained in the functional list object.
}
\description{
Unlike base R, this subsetting function uses cardinal instead of ordinal
numbers. This means that counting starts from 0.
}

