% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{position}
\alias{position}
\title{Return the position of the first object in a list that fulfills a predicate}
\usage{
position(ls, fn, start = 0)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{fn}{A predicate function, as described above}

\item{start}{A index of the list to begin the search (remember, cardinal numbers)}
}
\value{
Either the first object that fulfills the predicate or \code{FALSE}.
}
\description{
A predicate function is one that takes a single argument returns either true or
false. \code{position} takes such a function, applies it to a functional list,
and returns the position of first object where the function returns \code{TRUE}.
}
\details{
Remember, in this package we use cardinal numbers, counting from 0.
}
\seealso{
\code{\link[base]{funprog}}
}

