% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepend.R
\name{prepend}
\alias{prepend}
\title{Prepend adds an element to the list through recursion.}
\usage{
prepend(ls, el)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{el}{A data element (from R's standard data types) to add to the list}
}
\value{
A closure with the \code{fl} class added. The closure can be used to access
different properties of the list: either the head, tail or its empty status.
}
\description{
This is the primary list building tool. Although the user might usually prefer
to work with \code{\link{append}}, it is incredibly useful to use \code{prepend}
in other functions that work on functional lists. In fact, we need this function
to implement \code{\link{append}}.
}

