% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-fl.R
\name{print.fl}
\alias{print.fl}
\title{Print the contents of a functional list}
\usage{
\method{print}{fl}(x, ...)
}
\arguments{
\item{x}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{...}{Other arguments to print (for argument consistency)}
}
\description{
Print the contents of a functional list
}

