% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{reduce}
\alias{reduce}
\title{Reduce}
\usage{
reduce(ls, fn, ...)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{fn}{A binary function, i.e. a function that takes two arguments}

\item{...}{Additional arguments to \code{fn}}
}
\value{
A scalar
}
\description{
Recursively fold a function across a list, yielding a scalar.
}
\seealso{
\code{\link[base]{Reduce}}
}

