% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{remove.default}
\alias{remove}
\alias{remove.default}
\title{New Remove Generic}
\usage{
\method{remove}{default}(..., list = character(), pos = -1,
  envir = as.environment(pos), inherits = FALSE)

remove(...)
}
\arguments{
\item{...}{Arguments to different versions of \code{remove}}
}
\description{
So that we can have both the original and new version in our package
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Preserve original function
}}
\seealso{
\code{\link[base]{remove}}
}

