% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\name{rerun}
\alias{rerun}
\title{Repeatedly execute a quoted R expression}
\usage{
rerun(n, expr)
}
\arguments{
\item{n}{An integer that provides the desired number of executions}

\item{expr}{A quoted R expression to be iteratively executed}
}
\value{
A functional list of length \code{n}, where each element
represents one execution of \code{expr}
}
\description{
Repeatedly execute a quoted R expression
}
\examples{
rerun(10, quote(rnorm(1)))

}
\seealso{
\link[base]{replicate}
}

