% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-fl.R
\name{seq.fl}
\alias{seq.fl}
\title{Create a sequence of integers}
\usage{
\method{seq}{fl}(empty, start, end)
}
\arguments{
\item{empty}{An empty functional list, used to take advance of S3}

\item{start}{An integer specifying the begin of the sequence}

\item{end}{An integer that is the end of the sequence (but not included)}
}
\value{
A functional list that contains the sequence of integers
}
\description{
This function generates a functional list that generates a sequence of
integers in a manner similar to \code{range} in \code{Python}.
}

