% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{sequence}
\alias{sequence}
\title{Create a sequence of integers}
\usage{
sequence(start, end)
}
\arguments{
\item{start}{An integer specifying the begin of the sequence}

\item{end}{An integer that is the end of the sequence (but not included)}
}
\value{
A functional list that contains the sequence of integers
}
\description{
This function generates a functional list that generates a sequence of
integers in a manner similar to \code{range} in \code{Python}.
}

