% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{tail.fl}
\alias{tail.fl}
\title{Access the tail component of a function list object}
\usage{
\method{tail}{fl}(ls)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}
}
\value{
The tail argument of a functional list
}
\description{
This function is part of the \link[=lambdaList]{List out of Lambda} set of functions.
Like \code{\link{head}} and \code{\link{is_empty}}, this uses an anoymous function to
return a particular component of a functional list. Here, that's the tail argument.
}

