% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{take}
\alias{take}
\title{Keep all elements up to but not including n}
\usage{
take(ls, n)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{n}{An integer referring the cardinal position of a scalar in the
functional list object}
}
\value{
A functional list containing all of the scalars before the \emph{nth}
position
}
\description{
Unlike base R, this subsetting function uses cardinal instead of ordinal
numbers. This means that counting starts from 0 and continues up to, but not
including the \emph{nth} number.
}

