% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{length.fl}
\alias{length.fl}
\title{Find the length of a list}
\usage{
\method{length}{fl}(ls)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}
}
\value{
A scalar equal to the number of scalars contained in the functional list.
}
\description{
Here, the length equals the number of scalar data objects that it contations.
}

