% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaList.R
\name{remove}
\alias{remove}
\title{Return values that don't fulfill a predicate function}
\usage{
remove(ls, fn)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{fn}{A predicate function, as described above}
}
\value{
A functional list whose data elements return \code{false} in the predicate
function.
}
\description{
A predicate function is one that takes a single argument returns either true or
false. \code{remove} takes such a function, applies it to a functional list,
and keeps those values that return false in the predicate.
}
\seealso{
\code{\link{filter}}, \code{\link[base]{Filter}}
}

