% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\title{Select all of the elements between two positions in a list}
\usage{
slice(ls, start, end)
}
\arguments{
\item{ls}{A function, with \emph{class} \code{fl}, that operates as a functional
list object}

\item{start}{An integer referring the cardinal position of the first element in the slice}

\item{end}{An integer referring to the position after the final element in the slice}
}
\value{
A functional list containing all of the scalars from the \code{start}
position but before the \code{end} position
}
\description{
Unlike base R, this subsetting function uses cardinal instead of ordinal
numbers. This means that counting starts from 0 and continues up to, but not
including the \emph{nth} number.
}

